pipeline/requirements.txt
httpx==0.28.*
openai==1.82.*
pdfplumber==0.11.*
pytesseract==0.3.*
Pillow==11.*
celery[redis]==5.4.*
redis==5.2.*
supabase==2.13.*
psycopg2-binary==2.9.*
qdrant-client==1.13.*
python-dotenv==1.1.*
tenacity==9.*
tqdm==4.*
